Improve structure and documentation of Number Range connector
